move DatePickerInstance to instance.ts
