+++ packages/components/date-picker/index.ts
@@ -1,9 +1,7 @@
 import DatePicker from './src/date-picker'
 
-import type { App, ComponentPublicInstance } from 'vue'
+import type { App } from 'vue'
 import type { SFCWithInstall } from '@element-plus/utils'
-import type { DatePickerExpose } from './src/date-picker.type'
-import type { DatePickerProps } from './src/props/date-picker'
 
 const _DatePicker = DatePicker as SFCWithInstall<typeof DatePicker>
 
@@ -15,8 +13,4 @@ export default _DatePicker
 export const ElDatePicker = _DatePicker
 export * from './src/constants'
 export * from './src/props/date-picker'
-
-export type DatePickerInstance = ComponentPublicInstance<
-  DatePickerProps,
-  DatePickerExpose
->
+export type { DatePickerInstance } from './src/instance'
